left panel: reorganise files
