+++ src/app/ui/left-panel/pdbe-left-panel.tsx
@@ -1,0 +1,17 @@
+import { LeftPanel } from './core';
+import { CoreLeftPanelTabs, PDBeLeftPanelTabs } from './tabs';
+
+
+export const DefaultLeftPanelControls = LeftPanel([
+    CoreLeftPanelTabs.root,
+    CoreLeftPanelTabs.data,
+    CoreLeftPanelTabs.states,
+    CoreLeftPanelTabs.help,
+    PDBeLeftPanelTabs.segments,
+    CoreLeftPanelTabs.settings,
+], { defaultTab: 'none' });
+
+
+// TODO: export const PDBeLeftPanelControls = LeftPanel([...]);
+
+// TODO: create tab registry and allow only using tab names
